backend/requirements.txt
# AI English Learning System - Requirements
# FastAPI Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-socketio>=5.11.0

# Data Validation
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.0

# AI Services
faster-whisper>=0.10.0
edge-tts>=6.1.0

# HTTP Client
httpx>=0.26.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.2.0

# Logging
loguru>=0.7.0